move index finder up
